requirements.txt
#streamlit
gradio
google-generativeai
python-dotenv
pdf2image
